requirements.txt
azure-functions==1.0.5
azure-iot-device==2.0.0
azure-cognitiveservices-vision-computervision==0.5.0
requests==2.22.0
azure-cosmos==4.0.0b6
redis==3.3.11
markdown2==2.3.8
textgenrnn==1.5.0
tensorflow==2.0.0
twython==3.7.0
jsonschema==3.2.0